+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.1 (24/05/2025)
+- Improve the macOS codesign to add the `--deep` flag
+- Add `AssemblyVersionString` property
+- Transform all `EntryApplication` members into properties
+
 # 1.1.0 (18/05/2025)
 - Rename the `Updatum` class to `UpdatumManager` to simplify the name and avoid the requirement for fully qualifying name (#2)
 - Add a new constructor to pass the repository url, if null or empty it will try to infer from the RepositoryUrl attribute (#3)